add gpa calculation

+++ GPAcalculator/GPA_Calculator/Migrations/20230619101854_Initial.cs
@@ -49,7 +49,7 @@ namespace GPA_Calculator.Migrations
                     Id = table.Column<int>(type: "int", nullable: false)
                         .Annotation("SqlServer:Identity", "1, 1"),
                     SubjectName = table.Column<string>(type: "nvarchar(max)", nullable: false),
-                    Credits = table.Column<int>(type: "int", nullable: false)
+                    Credits = table.Column<double>(type: "float", nullable: false)
                 },
                 constraints: table =>
                 {
