+++ GPAcalculator/GPA_Calculator/Repositories/Dtos/RegisterSubject.cs
@@ -4,6 +4,6 @@
     {
        // public int StudentId { get; set; }
         public string SubjectName { get; set; }
-        public int Credits { get; set; }
+        public double Credits { get; set; }
     }
 }
